修改接口

+++ src/main/java/com/microfian/prac/response/ResWish.java
@@ -93,7 +93,7 @@ public class ResWish {
 
     private Integer pageSize;
 
-    private Integer apartDays;
+    private Long apartDays;
